Refactored LeaderboardEntry and ScoreStrategy to their own folder for better file structure
