+++ Backend/src/main/java/com/Pubrunda/controllers/LeaderboardEntryController.java
@@ -1,7 +1,7 @@
 package com.Pubrunda.controllers;
 
 import com.Pubrunda.exception.ResourceNotFoundException;
-import com.Pubrunda.models.LeaderboardEntry;
+import com.Pubrunda.models.leaderboardEntry.LeaderboardEntry;
 import com.Pubrunda.repositories.LeaderboardEntryRepository;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
