+++ Backend/src/main/java/com/Pubrunda/models/leaderboardEntry/LeaderboardEntry.java
@@ -1,12 +1,9 @@
-package com.Pubrunda.models;
+package com.Pubrunda.models.leaderboardEntry;
 
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 @Entity
 @Getter
