+++ Backend/src/main/java/com/Pubrunda/models/leaderboardEntry/ScoreStrategy.java
@@ -1,4 +1,4 @@
-package com.Pubrunda.models;
+package com.Pubrunda.models.leaderboardEntry;
 
 public interface ScoreStrategy {
