+++ Backend/src/main/java/com/Pubrunda/repositories/LeaderboardEntryRepository.java
@@ -1,6 +1,6 @@
 package com.Pubrunda.repositories;
 
-import com.Pubrunda.models.LeaderboardEntry;
+import com.Pubrunda.models.leaderboardEntry.LeaderboardEntry;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
